Complete implementation of Ticket #994 (Parallel Bjam output is jumbled).
I removed the signal handling code from jam.c as it's no longer needed.

Changed the execcmd function signature to pass two additional arguments,
the rule name and the build target.  These are no longer output in make1.c
as we need to defer their output until the command completes and we have
all the stdout and stderr output.

Fixed the various hangs and zombie children problems in execunix.


git-svn-id: http://svn.boost.org/svn/boost/branches/Boost_Jam_994@37855 b8fc166d-592f-0410-95f2-cb63ce0dd405