Rewrite much of the 'controlling the number of buckets' section.
I'm trying to make it clearer. It's a bit tricky as the standard doesn't guarantee much.
Instead of diving straight into the details I have tried to give the reader a rough
idea of what 'rehash' does and what the load factor is. This is hopefully enough to
understand the more detailled discussion of how you can control the number of buckets.
Then finally I discuss iterator invalidation.


git-svn-id: http://svn.boost.org/svn/boost/branches/unordered@42345 b8fc166d-592f-0410-95f2-cb63ce0dd405