Change includes of <cmath> to <boost/config/no_tr1/config.hpp>.
Previously if Boost.TR1 was in the include path then including <cmath> pulls in all the new TR1 math functions, which in turn also requires linking to an external library.  With auto-linking support this requires that library to have been built and be present in the library search path, even if the actual library under use is header only.

Fixes #2392.


git-svn-id: http://svn.boost.org/svn/boost/trunk/boost/numeric/ublas@49254 b8fc166d-592f-0410-95f2-cb63ce0dd405